# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nex Club API
  version: 0.1.0
  summary: 内容服务公开接口与管理接口。未实现的路径返回 501，code 为 not_implemented。
servers:
  - url: /
paths:
...
  /api/v1/resources:
    get:
      operationId: listResources
      tags: [public]
      parameters:
        - $ref: "#/components/parameters/kind"
        - name: q
          in: query
          schema: { type: string, default: "" }
        - name: tag
          in: query
          description: 重复参数表示交集，最多 8 个
          schema: { type: string }
          style: form
          explode: true
        - name: sort
          in: query
          schema: { $ref: "#/components/schemas/Sort" }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: 资源页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourcePage" }
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/resources/{id}:
    get:
      operationId: getResource
      tags: [public]
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: 已发布详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceDetail" }
        default:
          $ref: "#/components/responses/Error"
  /api/v1/tags:
    get:
      operationId: listTags
      tags: [public]
      parameters:
        - $ref: "#/components/parameters/kind"
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 标签计数。计数只受 kind 和 q 影响。
          content:
            application/json:
              schema:
                type: object
                required: [tags]
                properties:
                  tags:
                    type: array
                    items: { $ref: "#/components/schemas/TagCount" }
        default:
          $ref: "#/components/responses/Error"
...
  /api/admin/resources:
    get:
      operationId: listAdminResources
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/kind"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/Status" }
        - name: q
          in: query
          schema: { type: string }
      responses:
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createResource
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        default:
          $ref: "#/components/responses/Error"
  /api/admin/resources/{id}:
    get:
      operationId: getAdminResource
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        default:
          $ref: "#/components/responses/Error"
    put:
      operationId: saveResource
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        default:
          $ref: "#/components/responses/Error"
...
  /api/admin/resources/{id}/preview:
    get:
      operationId: previewResource
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: 草稿预览，形状与公开详情相同
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceDetail" }
        default:
          $ref: "#/components/responses/Error"
  /api/admin/resources/{id}/publish:
    post:
      operationId: publishResource
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [edit_version, revision_id]
              properties:
                edit_version: { type: integer }
                revision_id: { type: string, format: uuid }
      responses:
        default:
          $ref: "#/components/responses/Error"
  /api/admin/resources/{id}/visibility:
    post:
      operationId: setResourceVisibility
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [edit_version, status, reason]
              properties:
                edit_version: { type: integer }
                status: { $ref: "#/components/schemas/Status" }
                reason: { type: string }
      responses:
        default:
          $ref: "#/components/responses/Error"
  /api/admin/tags:
    get:
      operationId: listAdminTags
      tags: [admin]
      responses:
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createTag
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        default:
          $ref: "#/components/responses/Error"
  /api/admin/tags/{id}/merge:
    post:
      operationId: mergeTag
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/csrf"
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    mcpBearer:
      type: http
      scheme: bearer
    adminSession:
      type: apiKey
      in: cookie
      name: nex_session
  parameters:
    kind:
      name: kind
      in: query
      required: true
      schema: { $ref: "#/components/schemas/Kind" }
    resourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    csrf:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    Error:
      description: 错误信封。field_errors 没有时为 []。
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [code, message, request_id, field_errors]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        field_errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string }
    Kind:
      type: string
      enum: [tool, tutorial, repo]
    Status:
      type: string
      enum: [draft, published, hidden, archived]
    Sort:
      type: string
      enum: [recommended, heat, latest, relevance]
    Pricing:
      type: string
      enum: [unknown, free, paid, freemium]
    Level:
      type: string
      enum: [unknown, beginner, advanced]
...
    TagDimension:
      type: string
      enum: [category, capability, audience, difficulty]
    Platform:
      type: string
      enum: [web, desktop, mobile, api, plugin]
    Deployment:
      type: string
      enum: [hosted, self_host, local]
    TagRef:
      type: object
      required: [id, name, slug, dimension]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        dimension: { $ref: "#/components/schemas/TagDimension" }
    Card:
      type: object
      required: [subtitle, meta, href, cta]
      properties:
        subtitle: { type: string }
        meta: { type: string }
        href: { type: [string, "null"] }
        cta: { type: string }
    ResourceCard:
      type: object
      required: [id, kind, slug, title, summary, cover_urls, cover_fallback_count, tags, primary_category, quality_score, first_published_at, content_updated_at, card]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/Kind" }
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        cover_urls:
          type: array
          items: { type: string }
        cover_fallback_count: { type: integer }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TagRef" }
        primary_category:
          oneOf:
            - $ref: "#/components/schemas/TagRef"
            - type: "null"
        quality_score: { type: integer, minimum: 0, maximum: 100 }
        first_published_at: { type: [string, "null"], format: date-time }
        content_updated_at: { type: string, format: date-time }
        card: { $ref: "#/components/schemas/Card" }
    ResourceDetail:
      allOf:
        - $ref: "#/components/schemas/ResourceCard"
        - type: object
          required: [aliases, body_markdown, recommendation_reason, details]
          properties:
            aliases:
              type: array
              items: { type: string }
            body_markdown: { type: [string, "null"] }
            recommendation_reason: { type: [string, "null"] }
            details:
              oneOf:
                - $ref: "#/components/schemas/ToolDetails"
                - $ref: "#/components/schemas/TutorialDetails"
                - $ref: "#/components/schemas/RepoDetails"
    ToolDetails:
      type: object
      required: [website_url, pricing, platforms, deployment]
      properties:
        website_url: { type: string }
        pricing: { $ref: "#/components/schemas/Pricing" }
        platforms:
          type: array
          items: { $ref: "#/components/schemas/Platform" }
        deployment:
          type: array
          items: { $ref: "#/components/schemas/Deployment" }
    TutorialDetails:
      type: object
      required: [level, minutes, steps]
      properties:
        level: { $ref: "#/components/schemas/Level" }
        minutes: { type: integer, minimum: 0 }
        steps:
          type: array
          items: { type: string }
        author: { type: string }
        source_url: { type: [string, "null"] }
        notes: { type: [string, "null"] }
    RepoDetails:
      type: object
      required: [github_repository_id, full_name, language, license, archived]
      properties:
        github_repository_id: { type: [string, "null"] }
        full_name: { type: string }
        language: { type: [string, "null"] }
        license: { type: [string, "null"] }
        archived: { type: boolean }
        last_activity_at: { type: [string, "null"], format: date-time }
    ResourcePage:
      type: object
      required: [items, next_cursor, has_more, total, effective_sort, ranking_version, ranking_computed_at, applied_query]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ResourceCard" }
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }
        total: { type: [integer, "null"] }
        effective_sort: { $ref: "#/components/schemas/Sort" }
        ranking_version: { type: [string, "null"] }
        ranking_computed_at: { type: [string, "null"], format: date-time }
        applied_query:
          type: object
          required: [kind, q, tags, sort]
          properties:
            kind: { $ref: "#/components/schemas/Kind" }
            q: { type: string }
            tags:
              type: array
              items: { type: string }
            sort: { $ref: "#/components/schemas/Sort" }
    TagCount:
      allOf:
        - $ref: "#/components/schemas/TagRef"
        - type: object
          required: [count]
          properties:
            count: { type: integer, minimum: 0 }
...